Made all FBOs 16F

+++ game/SceneMain/BlurContainer.cpp
@@ -3,7 +3,7 @@
 #include "Manager.hpp"
 
 BlurContainer::BlurContainer() : noBlurDepth(vec2ui(0), TextureFormat::DEPTH_COMPONENT32) {
-	noBlurColor0.loadEmpty(vec2ui(0), TextureFormat::RGBA8);
+	noBlurColor0.loadEmpty(vec2ui(0), TextureFormat::RGBA16F);
 	noBlurColor0.setFilter(GL_NEAREST, GL_NEAREST);
 	noBlurColor0.setWrap(GL_CLAMP_TO_EDGE);
 	noBlur = new RenderTarget(1.0f);
@@ -13,19 +13,19 @@ BlurContainer::BlurContainer() : noBlurDepth(vec2ui(0), TextureFormat::DEPTH_COM
 	float blurSize = 1;
 	float blurSizeDivisor = std::pow(2,blurSize);
 
-	blurMaskColor0.loadEmpty(vec2ui(0), TextureFormat::RGBA8);
+	blurMaskColor0.loadEmpty(vec2ui(0), TextureFormat::RGBA16F);
 	blurMaskColor0.setFilter(GL_LINEAR, GL_LINEAR);
 	blurMaskColor0.setWrap(GL_CLAMP_TO_EDGE);
 	blurMask = new RenderTarget(1.0f);
 	blurMask->setTexture(RenderTargetBase::COLOR0, &blurMaskColor0);
 
-	horitzontalBlurredColor0.loadEmpty(vec2ui(0), TextureFormat::RGBA8);
+	horitzontalBlurredColor0.loadEmpty(vec2ui(0), TextureFormat::RGBA16F);
 	horitzontalBlurredColor0.setFilter(GL_LINEAR, GL_LINEAR);
 	horitzontalBlurredColor0.setWrap(GL_CLAMP_TO_EDGE);
 	horitzontalBlurred = new RenderTarget(1.0f/blurSizeDivisor);
 	horitzontalBlurred->setTexture(RenderTargetBase::COLOR0, &horitzontalBlurredColor0);
 
-	blurredColor0.loadEmpty(vec2ui(0), TextureFormat::RGBA8);
+	blurredColor0.loadEmpty(vec2ui(0), TextureFormat::RGBA16F);
 	blurredColor0.setFilter(GL_LINEAR, GL_LINEAR);
 	blurredColor0.setWrap(GL_CLAMP_TO_EDGE);
 	blurred = new RenderTarget(1.0f/blurSizeDivisor);
